fix NullPointerException in krun shutdown hook

Former-commit-id: bee3f7f296726039939b8b9a2125bc221759e0d0 [formerly 1057749cfdfbcecf12c44b0bf8f7bd3c3f12255a]
Former-commit-id: ab974945b333d9889fd2a3e4f1036d195f29d2e6

+++ src/javasources/KTool/src/org/kframework/krun/Main.java
@@ -951,7 +951,9 @@ public class Main {
             Runtime.getRuntime().addShutdownHook(new Thread() {
                 public void run() {
                     try {
-                        deleteDirectory(new File(K.krunTempDir));
+                        if (K.krunTempDir != null) {
+                            deleteDirectory(new File(K.krunTempDir));
+                        }
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
